Refactor Universe
