+++ Vang/CMakeFiles.cmake
@@ -37,6 +37,10 @@ set(Vang_FILES
 	"./src/Vang/Utility/Time/Time.h"
 	"./src/Vang/Utility/Time/Time.cpp"
 
+	"./src/Vang/Items/Block.h"
+	"./src/Vang/Items/Block.cpp"
+	"./src/Vang/Items/BlockBlueprint.h"
+	"./src/Vang/Items/BlockBlueprint.cpp"
 	"./src/Vang/Items/Item.h"
 	"./src/Vang/Items/Item.cpp"
 	"./src/Vang/Items/ItemBlueprint.h"
@@ -81,19 +85,8 @@ set(Vang_FILES
 	"./src/Vang/GraphicsAPI/GraphicsOpenGL/Shaders/VertexData.h"
 	"./src/Vang/GraphicsAPI/GraphicsOpenGL/Shaders/VertexData.cpp"
 
-	"./src/Vang/Voxel/Block.h"
-	"./src/Vang/Voxel/Block.cpp"
-	"./src/Vang/Voxel/BlockBlueprint.h"
-	"./src/Vang/Voxel/BlockBlueprint.cpp"
 	"./src/Vang/Voxel/Chunk.h"
 	"./src/Vang/Voxel/Chunk.cpp"
-	"./src/Vang/Voxel/ChunkLoader.h"
-	"./src/Vang/Voxel/ChunkLoader.cpp"
-	"./src/Vang/Voxel/CubeChunkLoader.h"
-	"./src/Vang/Voxel/CubeChunkLoader.cpp"
-	"./src/Vang/Voxel/ChunkRenderer.h"
-	"./src/Vang/Voxel/CubeChunkRenderer.h"
-	"./src/Vang/Voxel/CubeChunkRenderer.cpp"
 	"./src/Vang/Voxel/Universe.h"
 	"./src/Vang/Voxel/Universe.cpp"
 	"./src/Vang/Voxel/UniverseManager.h"
