+++ Vang/src/Vang.h
@@ -18,6 +18,7 @@
 #endif
 
 #include "Vang/GraphicsAPI/Generic/GraphicsAPI.h"
+#include "Vang/Items/BlockBlueprint.h"
 #include "Vang/Items/ItemBlueprint.h"
 #include "Vang/Modding/ModManager.h"
 #include "Vang/Utility/BlueprintContainer/BlueprintContainer.h"
@@ -27,8 +28,6 @@
 #include "Vang/Utility/Layers/LayerStack.h"
 #include "Vang/Utility/Player/Player.h"
 #include "Vang/Utility/Time/Time.h"
-#include "Vang/Voxel/BlockBlueprint.h"
-#include "Vang/Voxel/ChunkLoader.h"
 #include "Vang/Voxel/Universe.h"
 #include "Vang/Window/Window.h"
 
@@ -47,9 +46,4 @@ namespace Vang {
 	Vang::Utility::Layers::LayerStack& getLayerStack();
 	Vang::Utility::Events::EventHandler& getEventHandler();
 	Vang::Input::InputCache& getInputCache();
-
-	namespace detail {
-		uint32_t addChunkLoader(Vang::Voxel::ChunkLoader& chunkLoader);
-		void removeChunkLoader(uint32_t chunkLoaderId);
-	}
 };
