+++ Vang/src/Vang/Utility/Player/Player.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "Vang/Utility/Camera/Camera.h"
-#include "Vang/Voxel/ChunkRenderer.h"
 
 namespace Vang::Objects {
