+++ Vang/src/Vang/Voxel/Chunk.h
@@ -11,10 +11,10 @@ namespace Vang::Voxel {
 		Chunk();
 		Chunk(const glm::ivec3& chunkPosition);
 		Chunk(int32_t chunkX, int32_t chunkY, int32_t chunkZ);
-		Chunk(const Chunk&)						 = delete;
-		Chunk(Chunk&&)							 = delete;
-		Chunk& operator=(const Chunk&)			 = default;
-		Chunk& operator=(Chunk&&)				 = default;
+		Chunk(const Chunk&)			   = delete;
+		Chunk(Chunk&&)				   = delete;
+		Chunk& operator=(const Chunk&) = default;
+		Chunk& operator=(Chunk&&)	   = default;
 
 		void setBlock(uint32_t x, uint32_t y, uint32_t z, uint32_t block);
 		void setBlock(const glm::uvec3& coords, uint32_t block);
@@ -25,7 +25,7 @@ namespace Vang::Voxel {
 
 	private:
 		glm::ivec3 m_chunkPosition{};
-		std::array<uint32_t, 4096> m_blocks{};
+		std::vector<uint32_t> m_blocks{4096};
 
 		void generateChunk();
 	};
