--- Vang/src/Vang/Voxel/ChunkRenderer.cpp
@@ -1,9 +1,0 @@
-#include "ChunkRenderer.h"
-
-namespace Vang::Voxel {
-
-	uint32_t ChunkRenderer::getTotalRenderedChunks() const {
-		return m_totalRenderedChunks;
-	}
-
-}
