--- Vang/src/Vang/Voxel/CubeChunkRenderer.cpp
@@ -1,3 +1,0 @@
-#include "CubeChunkRenderer.h"
-
-namespace Vang::Voxel {}
