--- Vang/src/Vang/Voxel/CubeChunkRenderer.h
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace Vang::Voxel {
-
-	class CubeChunkRenderer {};
-
-}
