+++ Vang/src/Vang/Voxel/UniverseManager.cpp
@@ -3,14 +3,14 @@
 
 namespace Vang::Voxel::UniverseManager {
 
-	static Universe s_currentUniverse;
+	static std::shared_ptr<Universe> s_currentUniverse;
 
-	const Universe& createUniverse(const std::string& seed) {
-		s_currentUniverse = Universe{seed};
+	const std::shared_ptr<Universe> createUniverse(const std::string& seed) {
+		s_currentUniverse = std::make_shared<Universe>(seed);
 		return s_currentUniverse;
 	}
 
-	const Universe& getCurrentUniverse() {
+	const std::shared_ptr<Universe> getCurrentUniverse() {
 		return s_currentUniverse;
 	}
